Added check that all datapoints exist

+++ experiments/scripts_execution/summarize_results.sh
@@ -1,6 +1,6 @@
-if [ "$#" -ne 7 ]; then
+if [ "$#" -ne 9 ]; then
     # to run a full summary, set all arguments to True
-    echo "Usage: bash summarize_results.sh <aggregate> <create_plots> <create_numerical> <create_tables> <compile_plots> <compile_numerical> <compile_tables>"
+    echo "Usage: bash summarize_results.sh <aggregate> <count_data> <create_plots> <create_numerical> <create_tables> <compile_results> <compile_plots> <compile_numerical> <compile_tables>"
     exit 1
 fi
 
@@ -8,11 +8,16 @@ if [ "$1" = "True" ]; then
     mkdir -p ../data/aggregated
     rm -f ../data/aggregated/concatenated.csv
     rm -f ../data/aggregated/aggregated.csv
+    rm -f ../data/aggregated/counted.csv
     python3 ../scripts_analysis/ConcatFiles.py ../data/runs ../data/aggregated/concatenated.csv
     python3 ../scripts_analysis/AggregateRuns.py ../data/aggregated/concatenated.csv ../data/aggregated/aggregated.csv ../data/aggregated/counted.csv
 fi
 
 if [ "$2" = "True" ]; then
+    python3 ../scripts_analysis/CountData.py ../data/aggregated/concatenated.csv
+fi
+
+if [ "$3" = "True" ]; then
     mkdir -p ../data/plots
     rm -f ../data/plots/*
     mus=("2" "5" "10" "25")
@@ -24,12 +29,12 @@ if [ "$2" = "True" ]; then
     done
 fi
 
-if [ "$3" = "True" ]; then 
+if [ "$4" = "True" ]; then 
     mkdir -p ../data/numerical
     python3 ../scripts_analysis/AnalyzeNumerical.py ../data/aggregated/counted.csv ../data/aggregated/aggregated.csv ../data/numerical
 fi
 
-if [ "$4" = "True" ]; then
+if [ "$5" = "True" ]; then
     mkdir -p ../data/tables
     rm -f ../data/tables/*
     div_thresholds=()
@@ -41,7 +46,9 @@ if [ "$4" = "True" ]; then
     done
 fi
 
-python3 ../scripts_analysis/CompileResults.py ../data/other/result_file.tex $5 ../data/plots $6 ../data/numerical $7 ../data/tables
-pdflatex -output-directory=../data/other -shell-escape ../data/other/result_file.tex
-rm -f ../data/other/result_file.aux ../data/other/result_file.log 
-rm -r svg-inkscape
+if [ "$6" = "True" ]; then
+    python3 ../scripts_analysis/CompileResults.py ../data/other/result_file.tex $7 ../data/plots $8 ../data/numerical $9 ../data/tables
+    pdflatex -output-directory=../data/other -shell-escape ../data/other/result_file.tex
+    rm -f ../data/other/result_file.aux ../data/other/result_file.log 
+    rm -r svg-inkscape
+fi
